django/requirements.txt
aiohttp==3.7.4.post0
alibabacloud-alidns20150109==2.0.2
alibabacloud-bssopenapi20171214==1.0.8
alibabacloud-cdn20180510==1.0.8
alibabacloud-credentials==0.2.0
alibabacloud-dcdn20180115==1.0.2
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-live20161101==1.0.2
alibabacloud-openapi-util==0.1.6
alibabacloud-tea==0.2.8
alibabacloud-tea-openapi==0.3.1
alibabacloud-tea-util==0.3.5
aliyun-python-sdk-cdn==3.7.1
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-live==3.9.14
ansi==0.2.0
asgiref==3.3.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==17.4.0
Automat==0.6.0
awscli==1.20.56
beautifulsoup4==4.6.0
boto3==1.18.56
botocore==1.21.56
bottle==0.12.19
Brotli==1.0.9
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==8.0.3
cmake==3.21.1.post1
colorama==0.3.7
colorlog==5.0.1
command-not-found==0.3
configobj==5.0.6
configparser==5.0.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crayons==0.4.0
cryptography==3.4.7
daemonize==2.5.0
dataclasses==0.8
decorator==4.4.2
deepdiff==5.2.3
deepmerge==0.3.0
distro==1.6.0
distro-info===0.18ubuntu0.18.04.1
dj-rest-auth==2.1.11
Django==3.1.7
django-cors-headers==3.8.0
django-debug-toolbar==3.2.2
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
dnspython==2.1.0
docutils==0.15.2
drf-access-policy==1.1.0
drf-yasg==1.20.0
dulwich==0.20.21
edgegrid-python==1.1.1
errbot==6.1.8
et-xmlfile==1.1.0
ffprobe-python==1.0.3
Flask==2.0.2
flatten-dict==0.3.0
flatten-json==0.1.13
glob2==0.7
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
gspread==3.6.0
gunicorn==20.1.0
html5lib==0.999999999
httplib2==0.19.1
hyperlink==17.3.1
idna==2.10
idna-ssl==1.1.0
importlib-metadata==4.8.3
incremental==16.10.1
inflection==0.5.1
iso8601==1.0.2
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
json2html==1.3.0
jsoncompare==0.1.2
jsondiff==1.2.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lxml==4.2.1
m3u8==1.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
mongoengine==0.23.0
multidict==5.1.0
multivolumefile==0.2.3
mysqlclient==2.0.3
ndg-httpsclient==0.5.1
nested-lookup==0.2.22
netifaces==0.10.4
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.5.3.56
openpyxl==3.0.7
ordered-set==4.0.2
packaging==20.9
PAM==0.4.2
pandas==1.1.5
pathlib2==2.3.5
Pillow==8.2.0
prometheus-client==0.9.0
protobuf==3.17.0
py==1.10.0
py7zr==0.17.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==0.5.0
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.14.1
pycryptodomex==3.14.1
pyee==8.2.2
Pygments==2.8.1
pygments-markdown-lexer==0.1.0.dev39
PyGObject==3.26.1
PyJWT==1.7.1
pymongo==3.11.3
PyMySQL==1.0.2
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyppmd==0.17.3
pyserial==3.4
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2021.1
pyxdg==0.25
PyYAML==3.12
pyzstd==0.15.2
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
retry==0.9.2
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
schedule==1.1.0
scikit-build==0.11.1
SecretStorage==2.3.1
selenium==3.141.0
service-identity==16.0.0
six==1.16.0
slack-bolt==1.11.2
slack-sdk==3.13.0
slackclient==1.3.2
slackeventsapi==3.0.1
sqlparse==0.4.1
ssh-import-id==5.7
systemd-python==234
tabulate==0.8.9
tencentcloud-sdk-python-intl-en==3.0.433
texttable==1.6.4
tld==0.12.5
Twisted==17.9.0
typing-extensions==3.7.4.3
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.4
validators==0.18.2
waitress==2.0.0
wcwidth==0.2.5
webdriver-manager==3.3.0
webencodings==0.5
WebOb==1.8.7
websocket-client==0.54.0
WebTest==2.0.35
Werkzeug==2.0.2
xlrd==2.0.1
xmltodict==0.12.0
yarl==1.6.3
zipp==3.4.1
zope.interface==4.3.2
